+++ molecool/tests/test_molecule.py
@@ -2,6 +2,7 @@
 Tests for the molecule module.
 """
 
+import numpy as np
 import pytest
 import molecool
 
@@ -14,3 +15,15 @@ def test_molecular_mass():
 
     assert pytest.approx(actual_mass, abs=1e-2) == calculated_mass
 
+def test_build_bond_list():
+    coordinates = np.array([
+        [1,1,1],
+        [2.4,1,1],
+        [-0.4,1,1],
+        [1,1,2.4],
+        [1,1,-0.4]
+    ])
+
+    bonds = molecool.build_bond_list(coordinates)
+
+    assert len(bonds) == 4
